Vertex button updated

+++ src/model/Vertex.java
@@ -1,28 +1,28 @@
 package model;
 
 public class Vertex {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
     private String id;
 
-    public Vertex(int x, int y) {
+    public Vertex(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
